Network rendering working! Rewritten network code + more.
Finally got the sockets to work... Cleaned up the networking code too since making it share parts only made it hard to read and they don't really share many parts anymore.

It's setup in a way to make it possible to play on two instances with yourself for debugging reasons.